requirements.txt
papersizes==0.6.1
peewee==3.13.3
pycairo==1.19.1